pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "no-time-to-train"
version = "0.1.0"
description = "No Time to Train project"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "numpy==2.1.1",
    "pillow==10.4.0",
    "opencv-python==4.10.0.84",
    "tqdm==4.66.5",
    "pyyaml==6.0.2",
    "packaging==24.1",
    "requests==2.32.3",
    "pytorch-lightning==2.1.0",
    "torchmetrics==1.4.3",
    "lightning-utilities==0.11.7",
    "hydra-core==1.3.2",
    "omegaconf==2.3.0",
    "jsonargparse[signatures]==4.33.2",
    "timm==0.4.12",
    "kornia==0.7.4",
    "scikit-image==0.24.0",
    "imageio==2.35.1",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "pandas==2.2.3",
    "matplotlib==3.9.2",
    "seaborn==0.13.2",
    "joblib==1.4.2",
    "pycocotools==2.0.8",
    "lvis==0.5.3",
    "fvcore==0.1.5.post20221221",
    "iopath==0.1.10",
    "submitit==1.5.2",
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "wandb==0.18.2",
    "rich==13.9.2",
    "segment-anything==1.0",
    "addict==2.4.0",
    "termcolor==2.4.0",
    "tabulate==0.9.0",
    "yacs==0.1.8",
    "portalocker==2.10.1",
    "psutil==5.9.8",
    "cython==3.0.11",
    "pot==0.9.5",
    "mmengine==0.10.5",
    "nvitop==1.3.2",
    "nvidia-ml-py==12.535.161",
    "tidecv==1.0.1",
    "torch==2.4.1",
    "torchvision==0.19.1",
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "jupyterlab",
    "ipykernel",
    "ipywidgets"
]

[tool.setuptools.packages.find]
include = ["no_time_to_train*", "sam2*", "sam2_configs*"]
